build test for <App/>
